chore: Move hog function code to its own folder (#32120)
