+++ frontend/src/lib/components/ActivityLog/activityLogLogic.tsx
@@ -17,9 +17,9 @@ import { dataWarehouseSavedQueryActivityDescriber } from 'scenes/data-warehouse/
 import { errorTrackingActivityDescriber } from 'scenes/error-tracking/errorTrackingActivityDescriber'
 import { flagActivityDescriber } from 'scenes/feature-flags/activityDescriptions'
 import { groupActivityDescriber } from 'scenes/groups/activityDescriptions'
+import { hogFunctionActivityDescriber } from 'scenes/hog-functions/misc/activityDescriptions'
 import { notebookActivityDescriber } from 'scenes/notebooks/Notebook/notebookActivityDescriber'
 import { personActivityDescriber } from 'scenes/persons/activityDescriptions'
-import { hogFunctionActivityDescriber } from 'scenes/pipeline/hogfunctions/activityDescriptions'
 import { pluginActivityDescriber } from 'scenes/pipeline/pipelinePluginActivityDescriptions'
 import { insightActivityDescriber } from 'scenes/saved-insights/activityDescriptions'
 import { surveyActivityDescriber } from 'scenes/surveys/surveyActivityDescriber'
