+++ frontend/src/scenes/billing/ExportsUnsubscribeTable/exportsUnsubscribeTableLogic.tsx
@@ -5,7 +5,7 @@ import api from 'lib/api'
 import { getCurrentTeamId } from 'lib/utils/getAppContext'
 import { DESTINATION_TYPES } from 'scenes/pipeline/destinations/constants'
 import { pipelineDestinationsLogic } from 'scenes/pipeline/destinations/destinationsLogic'
-import { HogFunctionIcon } from 'scenes/pipeline/hogfunctions/HogFunctionIcon'
+import { HogFunctionIcon } from 'scenes/hog-functions/configuration/HogFunctionIcon'
 import { pipelineAccessLogic } from 'scenes/pipeline/pipelineAccessLogic'
 import { FunctionDestination, PipelineBackend } from 'scenes/pipeline/types'
 import { urls } from 'scenes/urls'
