+++ frontend/src/scenes/error-tracking/configuration/alerting/ErrorTrackingAlerting.tsx
@@ -1,5 +1,5 @@
-import { LinkedHogFunctions } from 'scenes/pipeline/hogfunctions/list/LinkedHogFunctions'
-import { HOG_FUNCTION_SUB_TEMPLATE_COMMON_PROPERTIES } from 'scenes/pipeline/hogfunctions/sub-templates/sub-templates'
+import { LinkedHogFunctions } from 'scenes/hog-functions/list/LinkedHogFunctions'
+import { HOG_FUNCTION_SUB_TEMPLATE_COMMON_PROPERTIES } from 'scenes/hog-functions/sub-templates/sub-templates'
 
 export function ErrorTrackingAlerting(): JSX.Element {
     return (
