+++ frontend/src/scenes/hog-functions/HogFunctionScene.tsx
@@ -5,14 +5,14 @@ import { NotFound } from 'lib/components/NotFound'
 import { LemonSkeleton } from 'lib/lemon-ui/LemonSkeleton'
 import { LemonTab, LemonTabs } from 'lib/lemon-ui/LemonTabs'
 import { capitalizeFirstLetter } from 'lib/utils'
-import { HogFunctionMetrics } from 'scenes/hog-functions/metrics/HogFunctionMetrics'
-import { HogFunctionTesting } from 'scenes/hog-functions/testing/HogFunctionTesting'
-import { HogFunctionConfiguration } from 'scenes/pipeline/hogfunctions/HogFunctionConfiguration'
+import { HogFunctionConfiguration } from 'scenes/hog-functions/configuration/HogFunctionConfiguration'
 import {
     hogFunctionConfigurationLogic,
     HogFunctionConfigurationLogicProps,
-} from 'scenes/pipeline/hogfunctions/hogFunctionConfigurationLogic'
-import { HogFunctionLogs } from 'scenes/pipeline/hogfunctions/logs/HogFunctionLogs'
+} from 'scenes/hog-functions/configuration/hogFunctionConfigurationLogic'
+import { HogFunctionLogs } from 'scenes/hog-functions/logs/HogFunctionLogs'
+import { HogFunctionMetrics } from 'scenes/hog-functions/metrics/HogFunctionMetrics'
+import { HogFunctionTesting } from 'scenes/hog-functions/testing/HogFunctionTesting'
 import { Scene, SceneExport } from 'scenes/sceneTypes'
 import { urls } from 'scenes/urls'
