+++ frontend/src/scenes/hog-functions/logs/HogFunctionLogs.tsx
