+++ frontend/src/scenes/hog-functions/logs/LogsViewer.tsx
