+++ frontend/src/scenes/hog-functions/logs/logsViewerLogic.tsx
