+++ frontend/src/scenes/hog-functions/metrics/HogFunctionEventEstimates.tsx
@@ -9,7 +9,7 @@ import { Query } from '~/queries/Query/Query'
 import { DataTableNode } from '~/queries/schema/schema-general'
 import { InsightType } from '~/types'
 
-import { hogFunctionConfigurationLogic } from '../hogFunctionConfigurationLogic'
+import { hogFunctionConfigurationLogic } from '../configuration/hogFunctionConfigurationLogic'
 const EVENT_THRESHOLD_ALERT_LEVEL = 8000
 
 export function HogFunctionEventEstimates(): JSX.Element | null {
