+++ frontend/src/scenes/hog-functions/misc/activityDescriptions.tsx
