+++ frontend/src/scenes/pipeline/destinations/destinationsLogic.tsx
@@ -6,6 +6,8 @@ import api from 'lib/api'
 import { FEATURE_FLAGS } from 'lib/constants'
 import { featureFlagLogic } from 'lib/logic/featureFlagLogic'
 import { deleteWithUndo } from 'lib/utils/deleteWithUndo'
+import { shouldShowHogFunction } from 'scenes/hog-functions/list/hogFunctionListLogic'
+import { hogFunctionTypeToPipelineStage } from 'scenes/hog-functions/misc/urls'
 import { projectLogic } from 'scenes/projectLogic'
 import { teamLogic } from 'scenes/teamLogic'
 import { userLogic } from 'scenes/userLogic'
@@ -20,8 +22,6 @@ import {
     PluginType,
 } from '~/types'
 
-import { shouldShowHogFunction } from '../hogfunctions/list/hogFunctionListLogic'
-import { hogFunctionTypeToPipelineStage } from '../hogfunctions/urls'
 import { pipelineAccessLogic } from '../pipelineAccessLogic'
 import {
     BatchExportDestination,
