+++ frontend/src/scenes/pipeline/destinations/newDestinationsLogic.tsx
@@ -5,15 +5,15 @@ import { combineUrl, router } from 'kea-router'
 import api from 'lib/api'
 import { FEATURE_FLAGS } from 'lib/constants'
 import { featureFlagLogic } from 'lib/logic/featureFlagLogic'
+import { HogFunctionIcon } from 'scenes/hog-functions/configuration/HogFunctionIcon'
+import { shouldShowHogFunctionTemplate } from 'scenes/hog-functions/list/hogFunctionTemplateListLogic'
+import { hogFunctionTypeToPipelineStage } from 'scenes/hog-functions/misc/urls'
 import { urls } from 'scenes/urls'
 import { userLogic } from 'scenes/userLogic'
 
 import { BATCH_EXPORT_SERVICE_NAMES, BatchExportService, HogFunctionTemplateType, PipelineStage } from '~/types'
 
 import { humanizeBatchExportName } from '../batch-exports/utils'
-import { HogFunctionIcon } from '../hogfunctions/HogFunctionIcon'
-import { shouldShowHogFunctionTemplate } from '../hogfunctions/list/hogFunctionTemplateListLogic'
-import { hogFunctionTypeToPipelineStage } from '../hogfunctions/urls'
 import { NewDestinationItemType, PipelineBackend } from '../types'
 import { RenderBatchExportIcon } from '../utils'
 import { destinationsFiltersLogic } from './destinationsFiltersLogic'
