+++ products/messaging/frontend/FunctionsTable.tsx
@@ -4,8 +4,8 @@ import { More } from 'lib/lemon-ui/LemonButton/More'
 import { LemonMenuOverlay } from 'lib/lemon-ui/LemonMenu/LemonMenu'
 import { updatedAtColumn } from 'lib/lemon-ui/LemonTable/columnUtils'
 import { LemonTableLink } from 'lib/lemon-ui/LemonTable/LemonTableLink'
-import { HogFunctionIcon } from 'scenes/pipeline/hogfunctions/HogFunctionIcon'
-import { HogFunctionStatusIndicator } from 'scenes/pipeline/hogfunctions/HogFunctionStatusIndicator'
+import { HogFunctionIcon } from 'scenes/hog-functions/configuration/HogFunctionIcon'
+import { HogFunctionStatusIndicator } from 'scenes/hog-functions/misc/HogFunctionStatusIndicator'
 import { urls } from 'scenes/urls'
 
 import { HogFunctionKind, HogFunctionType, HogFunctionTypeType } from '~/types'
